+++ apps/medon-fe/src/app/app.tsx
@@ -7,7 +7,6 @@ import { Navigate, Route, Routes } from 'react-router-dom';
 function App() {
   return (
     <Routes>
-      
       <Route index path="/" element={<Login />} />
       <Route path="/register" element={<RegistrationPage />} />
       <Route path="/forget-password" element={<ForgetPassword />} />
