+++ apps/medon-fe/src/redux/store.ts
@@ -1,17 +1,17 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { setupListeners } from '@reduxjs/toolkit/query';
 
-import counterReducer from './features/counter/counterSlice';
+import counterReducer from 'redux/features/fetchData/couter/counterSlice';
 import { fetchDataApi } from './features/fetchData/fetchData';
 
 export const store = configureStore({
-    reducer: {
-        counter: counterReducer,
-        [fetchDataApi.reducerPath]: fetchDataApi.reducer,
-    },
-    middleware: (getDefaultMiddleware) =>
-        getDefaultMiddleware().concat(fetchDataApi.middleware),
-    devTools: process.env.NODE_ENV !== 'production',
+  reducer: {
+    counter: counterReducer,
+    [fetchDataApi.reducerPath]: fetchDataApi.reducer,
+  },
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware().concat(fetchDataApi.middleware),
+  devTools: process.env.NODE_ENV !== 'production',
 });
 
 setupListeners(store.dispatch);
