+++ tsconfig.base.json
@@ -23,8 +23,8 @@
       "redux/*": ["apps/medon-fe/src/redux/*"],
       "utils/*": ["apps/medon-fe/src/utils/*"],
       "validation/*": ["apps/medon-fe/src/validation/*"],
-      "styles/*": ["apps/medon-fe/src/styles/*"],
-      "translation/*": ["apps/medon-fe/src/translation/*"]
+      "translation/*": ["apps/medon-fe/src/translation/*"],
+      "styles/*": ["apps/medon-fe/src/styles/*"]
     }
   },
   "exclude": ["node_modules", "tmp"]
